tests/wpt/web-platform-tests/workers/constructors/SharedWorker/no-arguments-ctor.html

<!--
/*
-->
<!doctype html>
<title>no arguments</title>
<link rel=help href="http://www.whatwg.org/html/#sharedworker">
<link rel=help href="http://dev.w3.org/2006/webapi/WebIDL/#es-interface-call">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws(new TypeError(), function() {
    var worker = new SharedWorker();
  });
});
</script>
<!--
*/
//-->
